// FrontEnd/src/app/model/Customer.ts
export class Customer {
 
  name: string;
  gender: string;
  address: string;
  city: string;
  state: string;
  country: string;
  hobbies: string;
  photo: string;
   
    // constructor(id:number, name:string, status:string) {
    //   this.gender = gender;
    //   this.name = name;
    //   this.status = status;

    //   name: string;
    //   gender: string;
    //   address: string;
    //   city: string;
    //   state: string;
    //   country: string;
    //   hobbies: string;
    //   photo: string;


    // }
     
  }